requirements.txt
langchain
langchain-community
langchain-groq
langchain-google-genai
langchain-experimental
pypdf
python-dotenv
pinecone-client
pinecone-client[grpc]
langchain-pinecone
flask
-e .